Remove logger member variable in TEB controller
Previously, the TEB controller contains a member variable "logger_".
This is bad because:
1. The member variable has to be inited inside the constructor of
the TEB controller, because the logger class is not default
constructable. On the other side, when the constructor is called,
the controoler has not yet configured. The lifetime managment is
messed up here.
2. If the controller first get a logger use "node->get_logger()",
and keep a private copy of it, and use the copy afterwards. This
usage assumes that each call of the function "get_logger" returns
an identical logger object. This is a very strong assumption, that
may break under some advanced logging cases, or maybe the internal
implemenation of "rcl::logger" is changed.