Fix runtime check
The current runtime check has the following problems:

1. TebAssertionFailureException should inherit from PlannerException
defined in "nav2_core/exceptions.hpp", because this is what the
controller server expects.

2. TEB_ASSERT_MSG uses a fixed lenght buffer on stack, and use sprintf
with it, this may cause buffer overflow.

3. Tempalate function is unnecessary, for it either dispatchs to integral
or pointer( ==0 check or ==nullptr check). And combination of macro and
template code is particularly bad (The TEB_ASSERT_MSG is a macro, expanded
to templates).

4. some values are not properly checked.