Fix some minor but annoying things
The following errors/inconsistencies are fixed

1. Clean some commented dead code

2. Delete the "initialized_" member variable of the controller class.
If the controller server tries to initilize the controller by
calling virtual function "configure", but the function
throws an exception, the controller server knows that the controller
is not properly initialized. You don't need a flag to indicate that
the object is initilized. That is already managed by lifetime
management of ROS2.

3. Since the "initialized_" flag is deleted, code related with
it are gone.

4. When init object of type "HomotopyClassPlanner" or
"TebOptimalPlanner", parameter "visualization_" should be nullptr.
An instance is created before calling the constructors
of the two.

5. The member variable of "map_frame" of TebConfig should be
defaulted to "map"

6. The member variable of "map_frame" of TebConfig should be inited
from "global_frame" of the YAML file. So make things consistent,
it is renamed as "global_frame"

7. The controller doesn't need a member variable "global_frame",
inited by "costmap_->getGlobalFrameID" and copy the value to
"cfg_->map_frame". We can just use the latter.

8. A configurable parameter "controller_frequency" is declared,
inited, intead hardwiring the value in the "configure" function.
Besides being configurable, the value should be defaulted to 5.0.