Fix static map loading problem
Previously, when TEB is brought up, the static map does not
show up. The reason is that in the configuration file for
RViz2, the duration policy for the map topic is mistakenly
set as "Volatile". As RViz2 starts up slowly, before it's
brought up, the map server has already published the static
map. Since the topic is published only once, and RViz2 fails
to get it, the map is not shown in RViz2. After the duration
policy is changed to "Transient Local", the problem
disappears.